+++ src/Support/Changelog.md
@@ -2,8 +2,8 @@
 
 ## 3.x Series
 
-## Unreleased
-##### 2023-XX-YY
+## 3.6.0
+##### 2023-03-07
 
 - Added Laravel 10 support
 - Added the `DetailedAmount` DTO class (provides a default implementation for the same name interface)